Refactor useEffects in BookingsTable and InvoicesTable into hooks.
